refactor: create module for conditions-formatting func
add extra_functions.js
